Minor modifications to the format.
Added a possible "sub" filter similar to PNG to improve LZ4 efficiency for raws.
Provide directly Android's Palette "most used swatches", instead of the list of most used colors.
The limit of 1024 colors for the palette was stupid, would need to encode the image as uint16..
All info structs are 4 bytes aligned.